Moving the navBar to layout folder
